+++ whitespace_format.py
@@ -3,7 +3,8 @@
 """Formatter of whitespace in text files.
 
 Author: David Pal <davidko.pal@gmail.com>
-Date: 2023
+Date: 2023 - 2024
+License: MIT License
 
 Usage:
 
@@ -13,21 +14,38 @@ Usage:
 from __future__ import annotations
 
 import argparse
-import copy
 import dataclasses
 import pathlib
 import re
 import sys
-from typing import Callable
-from typing import Dict
+from enum import Enum
 from typing import List
+from typing import Optional
+from typing import Tuple
 
 VERSION = "0.0.6"
 
 # Regular expression that does NOT match any string.
 UNMATCHABLE_REGEX = "$."
 
-END_OF_LINE_MARKERS = {
+# Whitespace characters
+CARRIAGE_RETURN = "\r"
+LINE_FEED = "\n"
+SPACE = " "
+TAB = "\t"
+VERTICAL_TAB = "\v"
+FORM_FEED = "\f"
+
+WHITESPACE_CHARACTERS = {
+    CARRIAGE_RETURN,
+    LINE_FEED,
+    SPACE,
+    TAB,
+    VERTICAL_TAB,
+    FORM_FEED,
+}
+
+NEW_LINE_MARKERS = {
     "windows": "\r\n",
     "linux": "\n",
     "mac": "\r",
@@ -58,6 +76,114 @@ COLORS = {
     "LIGHT_CYAN": "\033[96m",
     "WHITE": "\033[97m",
 }
+
+
+class ChangeType(Enum):
+    """Type of change that happened to a file."""
+
+    # New line marker was added to the end of the file (because it was missing).
+    NEW_LINE_MARKER_ADDED_TO_END_OF_FILE = 1
+
+    # New line marker was removed from the end of the file.
+    NEW_LINE_MARKER_REMOVED_FROM_END_OF_FILE = 2
+
+    # New line marker was replaced by another one.
+    REPLACED_NEW_LINE_MARKER = 3
+
+    # White at the end of a line was removed.
+    REMOVED_TRAILING_WHITESPACE = 4
+
+    # Empty line(s) at the end of file were removed.
+    REMOVED_EMPTY_LINES = 5
+
+    # An empty file was replaced by a file consisting of single empty line.
+    REPLACED_EMPTY_FILE_WITH_ONE_LINE = 6
+
+    # A file consisting of only whitespace was replaced by an empty file.
+    REPLACED_WHITESPACE_ONLY_FILE_WITH_EMPTY_FILE = 7
+
+    # A file consisting of only whitespace was replaced by a file consisting of single empty line.
+    REPLACED_WHITESPACE_ONLY_FILE_WITH_ONE_LINE = 8
+
+    # A tab character was replaces by space character(s).
+    REPLACED_TAB_WITH_SPACES = 9
+
+    # A tab character was removed.
+    REMOVED_TAB = 10
+
+    # A non-standard whitespace character (`\f` or `\v`) was replaced by a space character.
+    REPLACED_NONSTANDARD_WHITESPACE = 11
+
+    # A non-standard whitespace character (`\f` or `\v`) was removed.
+    REMOVED_NONSTANDARD_WHITESPACE = 12
+
+
+@dataclasses.dataclass
+class Change:
+    """Description of a change of the content of a file."""
+
+    change_type: ChangeType
+    line_number: int
+    changed_from: Optional[str] = None
+    changed_to: Optional[str] = None
+
+    def message(self, check_only: bool) -> str:
+        """Returns a message describing the change."""
+        check_only_word = "would be" if check_only else " "
+
+        if self.change_type == ChangeType.NEW_LINE_MARKER_ADDED_TO_END_OF_FILE:
+            return f"New line marker{check_only_word}added to the end of the file."
+
+        if self.change_type == ChangeType.NEW_LINE_MARKER_REMOVED_FROM_END_OF_FILE:
+            return f"New line marker{check_only_word}removed from the end of the file."
+
+        if self.change_type == ChangeType.REPLACED_NEW_LINE_MARKER:
+            return (
+                f"New line marker '{self.changed_from}'"
+                f"{check_only_word}replaced by '{self.changed_to}'."
+            )
+
+        if self.change_type == ChangeType.REMOVED_TRAILING_WHITESPACE:
+            return f"Trailing whitespace{check_only_word}removed."
+
+        if self.change_type == ChangeType.REMOVED_EMPTY_LINES:
+            return f"Empty line(s) at the end of the file{check_only_word}removed."
+
+        if self.change_type == ChangeType.REPLACED_EMPTY_FILE_WITH_ONE_LINE:
+            return f"Empty file{check_only_word}replaced with a single empty line."
+
+        if self.change_type == ChangeType.REPLACED_WHITESPACE_ONLY_FILE_WITH_EMPTY_FILE:
+            return f"File{check_only_word}replaced with an empty file."
+
+        if self.change_type == ChangeType.REPLACED_WHITESPACE_ONLY_FILE_WITH_ONE_LINE:
+            return f"File{check_only_word}replaced with a single empty line."
+
+        if self.change_type == ChangeType.REPLACED_TAB_WITH_SPACES:
+            return f"Tab{check_only_word}replaced with spaces."
+
+        if self.change_type == ChangeType.REMOVED_TAB:
+            return f"Tab{check_only_word}removed."
+
+        if self.change_type == ChangeType.REPLACED_NONSTANDARD_WHITESPACE:
+            return (
+                f"Non-standard whitespace character '{self.changed_from}'"
+                f"{check_only_word}replaced by a space."
+            )
+
+        if self.change_type == ChangeType.REMOVED_NONSTANDARD_WHITESPACE:
+            return (
+                f"Non-standard whitespace character '{self.changed_from}'{check_only_word}removed."
+            )
+
+        raise ValueError(f"Unknown change type: {self.change_type}")
+
+    def color_print(self, parsed_arguments: argparse.Namespace) -> None:
+        """Prints a message in color."""
+        color_print(
+            f"[BOLD][BLUE]↳ line {self.line_number + 1}: "
+            f"[WHITE]{self.message(parsed_arguments.check_only)}[RESET_ALL]",
+            parsed_arguments,
+        )
 
 
 def color_print(message: str, parsed_arguments: argparse.Namespace):
@@ -108,369 +234,293 @@ def write_file(file_name: str, file_content: str, encoding: str):
         die(4, f"Cannot write to file '{file_name}': {exception}")
 
 
-@dataclasses.dataclass
-class Line:
-    """Line of a text file.
-
-    The line is split into two parts:
-        1) Content
-        2) End of line marker ("\n", or "\r", or "\r\n")
-    """
-
-    content: str
-    end_of_line_marker: str
-
-    @staticmethod
-    def create_from_string(line: str) -> Line:
-        """Creates a line from a string.
-
-        The function splits the input into content and end_of_line_marker.
-        """
-        for end_of_line_marker in ["\r\n", "\n", "\r"]:
-            if line.endswith(end_of_line_marker):
-                return Line(line[: -len(end_of_line_marker)], end_of_line_marker)
-        return Line(line, "")
-
-    def to_hex(self):
-        """Returns a human-readable hexadecimal representation of the line.
-
-        This function is for debugging purposes only. It is used only during development.
-        """
-        return f"({string_to_hex(self.content)}, {string_to_hex(self.end_of_line_marker)})"
+def is_file_whitespace(file_content: str) -> bool:
+    """Determines if a file consists of only whitespace characters."""
+    for char in file_content:
+        if char not in WHITESPACE_CHARACTERS:
+            return False
+    return True
 
 
-def split_lines(text: str) -> List[Line]:
-    """Splits a string into lines."""
-    lines: List[Line] = []
-    current_line = ""
-    for i, char in enumerate(text):
-        current_line += char
-        if (char == "\n") or (
-            (char == "\r") and ((i >= len(text) - 1) or (not text[i + 1] == "\n"))
-        ):
-            lines.append(Line.create_from_string(current_line))
-            current_line = ""
-
-    if current_line:
-        lines.append(Line.create_from_string(current_line))
-
-    return lines
-
-
-def concatenate_lines(lines: List[Line]) -> str:
-    """Concatenates a list of lines into a single string including end-of-line markers."""
-    return "".join(line.content + line.end_of_line_marker for line in lines)
-
-
-def guess_end_of_line_marker(lines: List[Line]) -> str:
-    """Guesses the end of line marker.
+def find_most_common_new_line_marker(file_content: str) -> str:
+    """Finds the most common new line marker in a string.
 
     The function returns the most common end-of-line marker.
     Ties are broken in order Linux "\n", Mac "\r", Windows "\r\n".
     If no end-of-line marker is present, default to the Linux "\n" end-of-line marker.
     """
-    counts: Dict[str, int] = {"\n": 0, "\r": 0, "\r\n": 0}
-    for line in lines:
-        if line.end_of_line_marker in counts:
-            counts[line.end_of_line_marker] += 1
-    max_count = max(counts.values())
-    for end_of_line_marker, count in counts.items():
-        if count == max_count:
-            return end_of_line_marker
-    return "\n"  # This return statement is never executed.
+    linux_count = 0
+    macos_count = 0
+    windows_count = 0
+    i = 0
 
+    while i < len(file_content):
+        if file_content[i] == CARRIAGE_RETURN:
+            if i < len(file_content) - 1 and file_content[i + 1] == LINE_FEED:
+                windows_count += 1
+                i += 1
+            else:
+                macos_count += 1
+        elif file_content[i] == LINE_FEED:
+            linux_count += 1
+        i += 1
 
-def remove_trailing_empty_lines(lines: List[Line]) -> List[Line]:
-    """Removes trailing empty lines.
+    if macos_count > windows_count and macos_count > linux_count:
+        return "\r"
 
-    If there are no lines, empty list is returned.
-    If all lines are empty, the first line is kept.
-    """
-    num_empty_trailing_lines = 0
-    while (num_empty_trailing_lines < len(lines) - 1) and (
-        not lines[-num_empty_trailing_lines - 1].content
-    ):
-        num_empty_trailing_lines += 1
-    return copy.deepcopy(lines[: len(lines) - num_empty_trailing_lines])
+    if windows_count > linux_count:
+        return "\r\n"
 
-
-def remove_dummy_lines(lines: List[Line]) -> List[Line]:
-    """Remove empty lines that also have empty end-of-line markers."""
-    return [line for line in lines if line.content or line.end_of_line_marker]
-
-
-def remove_trailing_whitespace(lines: List[Line]) -> List[Line]:
-    """Removes trailing whitespace from every line."""
-    lines = [
-        Line(
-            re.sub(r"[ \n\r\t\f\v]*$", "", line.content),
-            line.end_of_line_marker,
-        )
-        for line in lines
-    ]
-    return remove_dummy_lines(lines)
-
-
-def normalize_end_of_line_markers(lines: List[Line], new_end_of_line_marker: str) -> List[Line]:
-    """Replaces end-of-line marker in all lines with a new end-of-line marker.
-
-    Lines without end-of-line markers (i.e. possibly the last line) are left unchanged.
-    """
-    return [
-        Line(line.content, new_end_of_line_marker) if line.end_of_line_marker else line
-        for line in lines
-    ]
-
-
-def remove_all_end_of_line_markers_from_end_of_file(lines: List[Line]) -> List[Line]:
-    """Removes all end-of-line markers from the end of the file."""
-    lines = remove_trailing_empty_lines(lines)
-    if not lines:
-        return []
-    lines[-1] = Line(lines[-1].content, "")
-    return remove_dummy_lines(lines)
-
-
-def add_end_of_line_marker_at_end_of_file(
-    lines: List[Line], new_end_of_line_marker: str
-) -> List[Line]:
-    """Adds new end-of-line marker to the end of file if it is missing."""
-    if not lines:
-        return [Line("", new_end_of_line_marker)]
-    lines = copy.deepcopy(lines)
-    lines[-1] = Line(lines[-1].content, new_end_of_line_marker)
-    return lines
-
-
-def normalize_empty_file(lines: List[Line], mode: str, new_end_of_line_marker: str) -> List[Line]:
-    """Replaces file with an empty file."""
-    if mode == "empty":
-        return []
-    if mode == "one-line":
-        return [Line("", new_end_of_line_marker)]
-    return copy.deepcopy(lines)
-
-
-def is_whitespace_only(lines: List[Line]) -> bool:
-    """Determines if file consists only of whitespace."""
-    for line in lines:
-        if line.content.strip(" \n\r\t\v\f"):
-            return False
-    return True
-
-
-def normalize_non_standard_whitespace(lines: List[Line], mode: str) -> List[Line]:
-    """Removes non-standard whitespace characters."""
-    if mode == "ignore":
-        return copy.deepcopy(lines)
-    if mode == "replace":
-        return [
-            Line(line.content.translate(str.maketrans("\v\f", "  ", "")), line.end_of_line_marker)
-            for line in lines
-        ]
-    return [
-        Line(line.content.translate(str.maketrans("", "", "\v\f")), line.end_of_line_marker)
-        for line in lines
-    ]
-
-
-def replace_tabs_with_spaces(lines: List[Line], num_spaces: int) -> List[Line]:
-    """Replaces tabs with spaces."""
-    if num_spaces < 0:
-        return copy.deepcopy(lines)
-    return [
-        Line(line.content.replace("\t", num_spaces * " "), line.end_of_line_marker)
-        for line in lines
-    ]
-
-
-def compute_difference(original_lines: List[Line], new_lines: List[Line]) -> List[int]:
-    """Computes the indices of lines that differ."""
-    line_numbers = [
-        line_number
-        for line_number, (original_line, new_line) in enumerate(zip(original_lines, new_lines))
-        if not original_line == new_line
-    ]
-    if len(original_lines) != len(new_lines):
-        line_numbers.append(min(len(original_lines), len(new_lines)))
-    return line_numbers
-
-
-@dataclasses.dataclass
-class ChangeDescription:
-    """Description of a change of the content of a file."""
-
-    check_only: str
-    change: str
-
-
-@dataclasses.dataclass
-class LineChange:
-    """Description of a change on a particular line."""
-
-    check_only: str
-    change: str
-    line_number: int
-
-
-class FileContentTracker:
-    """Tracks changes of the content of a file as it undergoes formatting."""
-
-    def __init__(self, lines: List[Line]):
-        """Initializes an instance of the file content tracker."""
-        self.initial_lines = lines
-        self.lines = copy.deepcopy(lines)
-        self.line_changes: List[LineChange] = []
-
-    def format(self, change: ChangeDescription, function: Callable[..., List[Line]], *args):
-        """Applies a change to the content of the file."""
-        previous_content = self.lines
-        self.lines = function(self.lines, *args)
-        if previous_content != self.lines:
-            line_numbers = compute_difference(previous_content, self.lines)
-            for line_number in line_numbers:
-                self.line_changes.append(LineChange(change.check_only, change.change, line_number))
-
-    def is_changed(self) -> bool:
-        """Determines if the file content has changed."""
-        return self.lines != self.initial_lines
+    return "\n"
 
 
 def format_file_content(
-    file_content_tracker: FileContentTracker,
-    parsed_arguments: argparse.Namespace,
-):
-    """Formats the content of file represented as a string."""
-    new_line_marker = END_OF_LINE_MARKERS.get(
+    file_content: str, parsed_arguments: argparse.Namespace
+) -> Tuple[str, List[Change]]:
+    """Applies a function to the content of a file.
+
+    Args:
+        file_content: Content of the file.
+        parsed_arguments: Parsed command line arguments.
+
+    Returns:
+        A pair consisting of the formatted file content and a list of changes.
+    """
+    output_new_line_marker = NEW_LINE_MARKERS.get(
         parsed_arguments.new_line_marker,
-        guess_end_of_line_marker(file_content_tracker.initial_lines),
+        find_most_common_new_line_marker(file_content),
     )
 
-    if is_whitespace_only(file_content_tracker.initial_lines):
-        changes = {
-            "ignore": ChangeDescription("", ""),
-            "empty": ChangeDescription(
-                check_only="File needs to be replaced with an empty file.",
-                change="File was replaced with an empty file.",
-            ),
-            "one-line": ChangeDescription(
-                check_only=(
-                    f"File must be replaced with a single-line empty line {repr(new_line_marker)}."
-                ),
-                change=(
-                    f"File was replaced with a single-line empty line {repr(new_line_marker)}."
-                ),
-            ),
-        }
-        if not file_content_tracker.initial_lines:
-            file_content_tracker.format(
-                changes[parsed_arguments.normalize_empty_files],
-                normalize_empty_file,
-                parsed_arguments.normalize_empty_files,
-                new_line_marker,
-            )
+    # Handle empty file:
+    if not file_content:
+        if parsed_arguments.normalize_empty_files in ["ignore", "empty"]:
+            return "", []
+        if parsed_arguments.normalize_empty_files == "one-line":
+            return output_new_line_marker, [Change(ChangeType.REPLACED_EMPTY_FILE_WITH_ONE_LINE, 1)]
+
+    # Handle non-empty file consisting of whitespace only.
+    if is_file_whitespace(file_content):
+        if parsed_arguments.normalize_whitespace_only_files == "empty":
+            return "", [Change(ChangeType.REPLACED_WHITESPACE_ONLY_FILE_WITH_EMPTY_FILE, 1)]
+        if parsed_arguments.normalize_whitespace_only_files == "one-line":
+            if file_content == output_new_line_marker:
+                return file_content, []
+            return output_new_line_marker, [
+                Change(ChangeType.REPLACED_WHITESPACE_ONLY_FILE_WITH_ONE_LINE, 1)
+            ]
+        if parsed_arguments.normalize_whitespace_only_files == "ignore":
+            return file_content, []
+
+    # Index into the input buffer.
+    i = 0
+
+    # List of changes
+    changes: List[Change] = []
+
+    # Line number. It is incremented every time we encounter a new end of line marker.
+    line_number = 1
+
+    # Position one byte past the end of last line in the output buffer
+    # excluding the last end of line marker.
+    last_end_of_line_excluding_eol_marker = 0
+
+    # Position one byte past the end of last line in the output buffer
+    # including the last end of line marker.
+    last_end_of_line_including_eol_marker = 0
+
+    # Position one byte past the last non-whitespace character in the output buffer.
+    last_non_whitespace = 0
+
+    # Position one byte past the end of last non-empty line in the output buffer
+    # excluding the last end of line marker.
+    last_end_of_non_empty_line_excluding_eol_marker = 0
+
+    # Position one byte past the end of last non-empty line in the output buffer,
+    # including the last end of line marker.
+    last_end_of_non_empty_line_including_eol_marker = 0
+
+    # Line number of the last non-empty line.
+    last_non_empty_line_number = 0
+
+    # Formatted output
+    output = ""
+
+    while i < len(file_content):
+        if file_content[i] in [CARRIAGE_RETURN, LINE_FEED]:
+            # Parse the new line marker
+            new_line_marker = ""
+            if file_content[i] == LINE_FEED:
+                new_line_marker = LINE_FEED
+            elif i < len(file_content) - 1 and file_content[i + 1] == LINE_FEED:
+                new_line_marker = "\r\n"
+                # Windows new line marker consists of two bytes.
+                # Skip the extra byte.
+                i += 1
+            else:
+                new_line_marker = CARRIAGE_RETURN
+
+            # Remove trailing whitespace
+            if parsed_arguments.remove_trailing_whitespace and max(
+                last_non_whitespace, last_end_of_line_including_eol_marker
+            ) < len(output):
+                changes.append(
+                    Change(
+                        ChangeType.REMOVED_TRAILING_WHITESPACE,
+                        line_number,
+                    )
+                )
+                output = output[
+                    : max(
+                        last_non_whitespace,
+                        last_end_of_line_including_eol_marker,
+                    )
+                ]
+
+            # Determine if the last line is empty
+            is_empty_line: bool = last_end_of_line_including_eol_marker == len(output)
+
+            # Add new line marker
+            last_end_of_line_excluding_eol_marker = len(output)
+            if (
+                parsed_arguments.normalize_new_line_markers
+                and output_new_line_marker != new_line_marker
+            ):
+                changes.append(
+                    Change(
+                        ChangeType.REPLACED_NEW_LINE_MARKER,
+                        line_number,
+                        new_line_marker,
+                        output_new_line_marker,
+                    )
+                )
+                output += output_new_line_marker
+            else:
+                output += new_line_marker
+
+            last_end_of_line_including_eol_marker = len(output)
+
+            # Update position of last non-empty line.
+            if not is_empty_line:
+                last_end_of_non_empty_line_excluding_eol_marker = (
+                    last_end_of_line_excluding_eol_marker
+                )
+                last_end_of_non_empty_line_including_eol_marker = (
+                    last_end_of_line_including_eol_marker
+                )
+                last_non_empty_line_number = line_number
+
+            line_number += 1
+
+        elif file_content[i] == SPACE:
+            output += file_content[i]
+
+        elif file_content[i] == TAB:
+            if parsed_arguments.replace_tabs_with_spaces < 0:
+                output += file_content[i]
+            elif parsed_arguments.replace_tabs_with_spaces > 0:
+                changes.append(Change(ChangeType.REPLACED_TAB_WITH_SPACES, line_number))
+                output += SPACE * parsed_arguments.replace_tabs_with_spaces
+            else:
+                # Remove the tab character.
+                changes.append(Change(ChangeType.REMOVED_TAB, line_number))
+
+        elif file_content[i] in [VERTICAL_TAB, FORM_FEED]:
+            if parsed_arguments.normalize_non_standard_whitespace == "ignore":
+                output += file_content[i]
+            elif parsed_arguments.normalize_non_standard_whitespace == "replace":
+                output += SPACE
+                changes.append(
+                    Change(
+                        ChangeType.REPLACED_NONSTANDARD_WHITESPACE,
+                        line_number,
+                        file_content[i],
+                        SPACE,
+                    )
+                )
+            elif parsed_arguments.normalize_non_standard_whitespace == "remove":
+                changes.append(
+                    Change(
+                        ChangeType.REMOVED_NONSTANDARD_WHITESPACE, line_number, file_content[i], ""
+                    )
+                )
+            else:
+                raise ValueError("Unknown value of normalize_non_standard_whitespace")
         else:
-            file_content_tracker.format(
-                changes[parsed_arguments.normalize_whitespace_only_files],
-                normalize_empty_file,
-                parsed_arguments.normalize_whitespace_only_files,
-                new_line_marker,
-            )
+            output += file_content[i]
+            last_non_whitespace = len(output)
 
-    else:
-        if parsed_arguments.remove_trailing_whitespace:
-            file_content_tracker.format(
-                ChangeDescription(
-                    check_only="Whitespace at the end of line needs to be removed.",
-                    change="Whitespace at the end of line was removed.",
-                ),
-                remove_trailing_whitespace,
-            )
+        # Move to the next character
+        i += 1
 
-        if parsed_arguments.remove_trailing_empty_lines:
-            file_content_tracker.format(
-                ChangeDescription(
-                    check_only="Empty line(s) at the end of file need to be removed.",
-                    change="Empty line(s) at the end of file were removed.",
-                ),
-                remove_trailing_empty_lines,
-            )
+    # Remove trailing whitespace from the last line.
+    if (
+        parsed_arguments.remove_trailing_whitespace
+        and last_end_of_line_including_eol_marker < len(output)
+        and last_non_whitespace < len(output)
+    ):
+        changes.append(Change(ChangeType.REMOVED_TRAILING_WHITESPACE, line_number))
+        output = output[:last_non_whitespace]
 
-        file_content_tracker.format(
-            ChangeDescription(
-                check_only="Tabs need to be replaced with spaces.",
-                change="Tabs were replaced by spaces.",
-            ),
-            replace_tabs_with_spaces,
-            parsed_arguments.replace_tabs_with_spaces,
-        )
+    # Remove trailing empty lines.
+    if (
+        parsed_arguments.remove_trailing_empty_lines
+        and last_end_of_line_including_eol_marker == len(output)
+        and last_end_of_non_empty_line_including_eol_marker < len(output)
+    ):
+        line_number = last_non_empty_line_number + 1
+        last_end_of_line_excluding_eol_marker = last_end_of_non_empty_line_excluding_eol_marker
+        last_end_of_line_including_eol_marker = last_end_of_non_empty_line_including_eol_marker
+        changes.append(Change(ChangeType.REMOVED_EMPTY_LINES, line_number))
+        output = output[:last_end_of_non_empty_line_including_eol_marker]
 
-        file_content_tracker.format(
-            ChangeDescription(
-                check_only=(
-                    "Non-standard whitespace characters need to be removed or replaced by spaces."
-                ),
-                change="Non-standard whitespace characters were removed or replaced by spaces.",
-            ),
-            normalize_non_standard_whitespace,
-            parsed_arguments.normalize_non_standard_whitespace,
-        )
+    # Add new line marker at the end of the file
+    if (
+        parsed_arguments.add_new_line_marker_at_end_of_file
+        and last_end_of_line_including_eol_marker < len(output)
+    ):
+        last_end_of_line_excluding_eol_marker = len(output)
+        changes.append(Change(ChangeType.NEW_LINE_MARKER_ADDED_TO_END_OF_FILE, line_number))
+        output += output_new_line_marker
+        last_end_of_line_including_eol_marker = len(output)
+        line_number += 1
 
-        if parsed_arguments.normalize_new_line_markers:
-            file_content_tracker.format(
-                ChangeDescription(
-                    check_only=(
-                        f"New line marker(s) need to be replaced with {repr(new_line_marker)}."
-                    ),
-                    change=f"New line marker(s) were replaced with {repr(new_line_marker)}.",
-                ),
-                normalize_end_of_line_markers,
-                new_line_marker,
-            )
+    # Remove new line marker from the end of the file
+    if (
+        parsed_arguments.remove_new_line_marker_from_end_of_file
+        and last_end_of_line_including_eol_marker == len(output)
+        and line_number >= 2
+    ):
+        line_number -= 1
+        changes.append(Change(ChangeType.NEW_LINE_MARKER_REMOVED_FROM_END_OF_FILE, line_number))
+        output = output[:last_end_of_line_excluding_eol_marker]
 
-        if parsed_arguments.add_new_line_marker_at_end_of_file:
-            file_content_tracker.format(
-                ChangeDescription(
-                    check_only=f"New line marker needs to be added to the end of the file, "
-                    f"or replaced with {repr(new_line_marker)}.",
-                    change=f"New line marker was added to the end of the file, "
-                    f"or replaced with {repr(new_line_marker)}.",
-                ),
-                add_end_of_line_marker_at_end_of_file,
-                new_line_marker,
-            )
-        elif parsed_arguments.remove_new_line_marker_from_end_of_file:
-            file_content_tracker.format(
-                ChangeDescription(
-                    check_only="New line marker(s) need to removed from the end of the file.",
-                    change="New line marker(s) were removed from the end of the file.",
-                ),
-                remove_all_end_of_line_markers_from_end_of_file,
-            )
+    return output, changes
 
 
 def reformat_file(file_name: str, parsed_arguments: argparse.Namespace) -> bool:
-    """Reformats a file."""
+    """Reformats a file.
+
+    Args:
+        file_name: Name of the file to reformat.
+        parsed_arguments: Parsed command line arguments.
+
+    Returns:
+         True if the file was changed, False otherwise.
+    """
     file_content = read_file_content(file_name, parsed_arguments.encoding)
-    lines = split_lines(file_content)
-    file_content_tracker = FileContentTracker(lines)
-    format_file_content(file_content_tracker, parsed_arguments)
-    is_changed = file_content_tracker.is_changed()
+    formatted_file_content, file_changes = format_file_content(file_content, parsed_arguments)
     if parsed_arguments.verbose:
         color_print(f"Processing file '{file_name}'...", parsed_arguments)
     if parsed_arguments.check_only:
-        if is_changed:
+        if file_changes:
             color_print(
                 f"[RED]✘[RESET_ALL] [BOLD][WHITE]{file_name} "
                 f"[RED]needs to be formatted[RESET_ALL]",
                 parsed_arguments,
             )
-            for line_change in file_content_tracker.line_changes:
-                color_print(
-                    f"   [BOLD][BLUE]↳ line {line_change.line_number + 1}: "
-                    f"[WHITE]{line_change.check_only}[RESET_ALL]",
-                    parsed_arguments,
-                )
+            for line_change in file_changes:
+                print("   ", end="")
+                line_change.color_print(parsed_arguments)
         else:
             if parsed_arguments.verbose:
                 color_print(
@@ -479,21 +529,21 @@ def reformat_file(file_name: str, parsed_arguments: argparse.Namespace) -> bool:
                     parsed_arguments,
                 )
     else:
-        if is_changed:
+        if file_changes:
             color_print(f"[WHITE]Reformatted [BOLD]{file_name}[RESET_ALL]", parsed_arguments)
-            for line_change in file_content_tracker.line_changes:
-                color_print(
-                    f"   [BOLD][BLUE]↳ line {line_change.line_number + 1}: "
-                    f"[WHITE]{line_change.change}[RESET_ALL]",
-                    parsed_arguments,
-                )
+            for line_change in file_changes:
+                print("   ", end="")
+                line_change.color_print(parsed_arguments)
             write_file(
-                file_name, concatenate_lines(file_content_tracker.lines), parsed_arguments.encoding
+                file_name,
+                formatted_file_content,
+                parsed_arguments.encoding,
             )
         else:
             if parsed_arguments.verbose:
                 color_print(f"[WHITE]{file_name} [BLUE]left unchanged[RESET_ALL]", parsed_arguments)
-    return is_changed
+
+    return bool(file_changes)
 
 
 def reformat_files(file_names: List[str], parsed_arguments: argparse.Namespace):
@@ -637,6 +687,9 @@ def parse_command_line() -> argparse.Namespace:
             "mac: Use Mac new line marker '\\r'. "
             "windows: Use Windows new line marker '\\r\\n'. "
         ),
+        type=str,
+        choices=["auto", "linux", "mac", "windows"],
+        default="auto",
     )
     parser.add_argument(
         "--normalize-new-line-markers",
@@ -685,7 +738,10 @@ def parse_command_line() -> argparse.Namespace:
     )
     parser.add_argument(
         "--remove-new-line-marker-from-end-of-file",
-        help="Remove new line markers from the end of each file.",
+        help="Remove new line markers from the end of each file. "
+        "This option conflicts with --add-new-line-marker-at-end-of-file. "
+        "This option implies --remove-trailing-empty-lines option, i.e., "
+        "all empty lines at the end of the file are removed.",
         required=False,
         default=False,
         action="store_true",
@@ -699,7 +755,8 @@ def parse_command_line() -> argparse.Namespace:
     )
     parser.add_argument(
         "--remove-trailing-empty-lines",
-        help="Remove empty lines at the end of each file.",
+        help="Remove empty lines at the end of each file. "
+        "If --remove-trailing-empty-lines is used, this option is used automatically.",
         required=False,
         default=False,
         action="store_true",
@@ -738,6 +795,16 @@ def parse_command_line() -> argparse.Namespace:
 
     if parsed_arguments.verbose:
         parsed_arguments.quiet = False
+
+    if parsed_arguments.remove_new_line_marker_from_end_of_file:
+        parsed_arguments.remove_empty_lines = True
+
+        if parsed_arguments.add_new_line_marker_from_end_of_file:
+            die(
+                1,
+                "The option --remove-new-line-marker-from-end-of-file "
+                "conflicts with --add-new-line-marker-at-end-of-file option.",
+            )
 
     return parsed_arguments
